- change type argument - specifying id generation strategy

+++ src/main/java/com/haotam/giupviec/repositories/WorkerRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.haotam.giupviec.entity.Worker;
 
-public interface WorkerRepository extends JpaRepository<Worker, Integer> {   
+public interface WorkerRepository extends JpaRepository<Worker, Long> {
 }
